backend/requirements.txt
blinker==1.9.0
cas==1.0.0
certifi==2025.1.31
charset-normalizer==2.0.12
click==8.1.8
cloudinary==1.34.0
Flask==3.1.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
google-generativeai==0.8.4
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==1.26.4
pinecone-client==2.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-cas==1.0.0
python-dotenv==0.19.0
requests==2.26.0
scikit-learn==1.3.2
scipy==1.12.0
setuptools==77.0.3
six==1.17.0
SQLAlchemy==2.0.39
threadpoolctl==3.6.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
